Create a User interceptor

+++ src/home/home.controller.ts
@@ -1,5 +1,6 @@
-import { Body, Controller, Delete, Get, Param, ParseIntPipe, Post, Put, Query } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, ParseIntPipe, Post, Put, Query, UnauthorizedException } from '@nestjs/common';
 import { PropertyType } from '@prisma/client';
+import { User, UserInfo } from 'src/user/decorators/user.decorator';
 import { CreateHomeDto, HomeResponseDto, UpdateHomeDto } from './dto/home.dto';
 import { HomeService } from './home.service';
 
@@ -34,17 +35,21 @@ export class HomeController {
     }
 
     @Post()
-    createHome(@Body() body: CreateHomeDto) {
-        return this.homeService.createHome(body);
+    createHome(@Body() body: CreateHomeDto, @User() user: UserInfo) {
+        return this.homeService.createHome(body, user.id);
     }
 
     @Put(':id')
-    updateHome(@Param('id', ParseIntPipe) id: number, @Body() body: UpdateHomeDto) {
+    async updateHome(@Param('id', ParseIntPipe) id: number, @Body() body: UpdateHomeDto, @User() user: UserInfo) {
+        const realtor = await this.homeService.getRealtorByHomeId(id);
+        if (realtor.id !== user.id) throw new UnauthorizedException();
         return this.homeService.updateHome(id, body);
     }
 
     @Delete(':id')
-    deleteHome(@Param('id', ParseIntPipe) id: number) {
+    async deleteHome(@Param('id', ParseIntPipe) id: number, @User() user: UserInfo) {
+        const realtor = await this.homeService.getRealtorByHomeId(id);
+        if (realtor.id !== user.id) throw new UnauthorizedException();
         return this.homeService.deleteHomeById(id);
     }
